requirements.txt
Django~=4.0.5
redis~=4.3.3
python-dotenv~=0.20.0